backend/requirements/common.txt
django==3.2
djangorestframework
markdown
django-filter
python-dotenv
django-cors-headers
django-allauth
pillow
djangorestframework-jwt